+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sensors/ray_caster/patterns/patterns_cfg.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 import torch
 from collections.abc import Callable, Sequence
 from dataclasses import MISSING
+from typing import Literal
 
 from omni.isaac.orbit.utils import configclass
 
@@ -33,16 +34,38 @@ class GridPatternCfg(PatternBaseCfg):
     """Configuration for the grid pattern for ray-casting.
 
     Defines a 2D grid of rays in the coordinates of the sensor.
+
+    .. attention::
+        The points are ordered based on the :attr:`ordering` attribute.
+
     """
 
     func: Callable = patterns.grid_pattern
 
     resolution: float = MISSING
     """Grid resolution (in meters)."""
+
     size: tuple[float, float] = MISSING
     """Grid size (length, width) (in meters)."""
+
     direction: tuple[float, float, float] = (0.0, 0.0, -1.0)
     """Ray direction. Defaults to (0.0, 0.0, -1.0)."""
+
+    ordering: Literal["xy", "yx"] = "xy"
+    """Specifies the ordering of points in the generated grid. Defaults to ``"xy"``.
+
+    Consider a grid pattern with points at :math:`(x, y)` where :math:`x` and :math:`y` are the grid indices.
+    The ordering of the points can be specified as "xy" or "yx". This determines the outer and inner loop order
+    when iterating over the grid points.
+
+    * If *"xy"* is selected, the points are ordered with outer loop over "x" and inner loop over "y".
+    * If *"yx"* is selected, the points are ordered with outer loop over "y" and inner loop over "x".
+
+    For example, the grid pattern points with :math:`X = (0, 1, 2)` and :math:`Y = (3, 4)`:
+
+    * *"xy"* ordering: :math:`[(0, 3), (0, 4), (1, 3), (1, 4), (2, 3), (2, 4)]`
+    * *"yx"* ordering: :math:`[(0, 3), (1, 3), (2, 3), (1, 4), (2, 4), (2, 4)]`
+    """
 
 
 @configclass